requirements.txt
psutil==5.9.0
wmi==1.5.1
flask==2.0.1
sqlalchemy==1.4.23
pywin32==306
GPUtil==1.4.0
requests==2.26.0
python-dotenv==0.19.0
apscheduler==3.8.1
pytz==2023.3